genre fix, split logic into files, added margin to pagination wrapper
