+++ src/JS/localStorage.js
@@ -1,0 +1,6 @@
+function getFromLocalStorage(key) {
+  const savedMovies = localStorage.getItem(key);
+  return JSON.parse(savedMovies);
+}
+
+export { getFromLocalStorage };
