+++ src/JS/main.js
@@ -13,10 +13,9 @@ function genres() {
   const fetchGenres = async () => {
   const response = await fetch(`https://api.themoviedb.org/3/genre/movie/list?api_key=${API_KEY}&language=uk-UA`);
   const genres = await response.json();
-  console.log(genres);
-  localStorage.setItem("genres", JSON.stringify(genres));
-    };
-    fetchGenres();
+  localStorage.setItem("genres", JSON.stringify(genres))
+  };
+  fetchGenres();
 }
 }
 export { genres };
@@ -43,7 +42,8 @@ export { fetchQueryFilm };
   
   // Рендер карточек
 function renderFilms(films) {
-
+  const savedGenres = localStorage.getItem("genres");
+  const parseGenres = JSON.parse(savedGenres);
   const markup = films.results.map(({title, poster_path, genre_ids, release_date, first_air_date,
     year = release_date || first_air_date || ' - ',
   }) => { 
@@ -58,14 +58,13 @@ function renderFilms(films) {
         genreOutput = genreName.map(element=> element.name);
        }
 
-      //console.log(genreOutput);
+      console.log(genreOutput);
           return `<li class="gallery__item">
             <a class="gallery__link" href="">
                 <img class="gallery__image" src="https://image.tmdb.org/t/p/w500${poster_path}" alt="" loading="lazy">
             </a>
             <div class="gallery__info">
                 <p class="gallery__title">${title}</p>
-
                 <p class="gallery__genre">${genreOutput.join(', ')}</p>     
                 <p class="gallery__year">${year.slice(0, 4)}</p>
 
@@ -78,7 +77,8 @@ function renderFilms(films) {
 }
 export { renderFilms };
 
+
 //Парсінг жанрів
-const savedGenres = localStorage.getItem("genres");
-const parseGenres = JSON.parse(savedGenres);
-console.log(parseGenres.genres);
+// const savedGenres = localStorage.getItem("genres");
+// const parseGenres = JSON.parse(savedGenres);
+// console.log(parseGenres.genres);
